Unify pnpm e2e and pnpm e2e:examples into one e2e config
They were the same concept — itx e2e against a deployed OS — split only
because the examples suite grew its own vitest config for a browser
project. Now one config (e2e/vitest.config.ts) with two projects:

- node: the engine e2e suites (e2e/vitest/**) AND the itx catalogue
  matrix (e2e/examples/*.e2e.test.ts), in one lane
- browser: the catalogue in a real browser (vitest browser mode)

`pnpm e2e` runs the whole suite; `pnpm e2e --project node` scopes to the
node lane. `pnpm e2e:examples` and e2e/examples/vitest.config.ts are
gone. Preview CI runs `pnpm e2e --project node` as a single vitest lane
(was two concurrent lanes), and the two run roots collapse to one
(os-e2e-), so the second artifact glob goes too.

Docs (README, testing, AGENTS, src/README, ci-preview-performance) and
knip updated to match.
