+++ apps/os/README.md
@@ -64,8 +64,8 @@ Run from `apps/os`.
 pnpm dev                 # local OS dev with Doppler-backed env (all workers in one workerd)
 pnpm typecheck           # TypeScript (includes route-tree freshness check)
 pnpm test                # unit tests
-pnpm e2e                 # real-worker e2e (itx e2e suites + preview smoke) against a live deployment
-pnpm e2e:examples        # the itx example matrix across all execution runtimes
+pnpm e2e                 # real-worker e2e against a live deployment (engine suites + itx example matrix)
+pnpm e2e --project node  # just the node lane (skip the browser example matrix)
 pnpm cli itx run --eval 'return await itx.whoami()'
                          # run an itx script against the deployment in your Doppler config
 pnpm cli claude-mcp      # open Claude against the OS MCP server in your local Doppler config
@@ -79,11 +79,12 @@ Use `pnpm run deploy`, not `pnpm deploy`; `deploy` is also a pnpm built-in.
 
 ## Running Real-Worker Tests
 
-The e2e lanes run against a real OS deployment, not the Workers Vitest pool:
-`pnpm e2e` (config `e2e/vitest.config.ts`: `e2e/vitest/**` plus itx
-suites in `e2e/vitest/**`) and `pnpm e2e:examples` (config
-`e2e/examples/vitest.config.ts`: the example matrix, including a browser
-runtime). Start the worker in one terminal, then run tests from another
+The e2e suite runs against a real OS deployment, not the Workers Vitest pool:
+`pnpm e2e` (config `e2e/vitest.config.ts`, one config with two projects — a
+`node` project covering the engine suites in `e2e/vitest/**` and the itx
+example matrix in `e2e/examples/**`, and a `browser` project running the
+matrix in a real browser). Start the worker in one terminal, then run tests
+from another
 through the matching Doppler config. For local dev configs, test helpers read
 `.alchemy/dev-server.json` to find the selected port; deployed configs get
 `APP_CONFIG_BASE_URL` from Doppler. All lanes, targets, and the canonical env
