+++ apps/os/e2e/AGENTS.md
@@ -26,11 +26,12 @@ How to target local dev / previews / prd and the canonical env vars
 (`APP_CONFIG_BASE_URL`, `APP_CONFIG_ADMIN_API_SECRET`, the `OS_E2E_*` harness knobs) are documented
 in [docs/testing.md](../../../docs/testing.md).
 
-- Live deployment tests: `pnpm e2e` (itx e2e suites in `e2e/vitest/`, agents, admin-project,
-  preview smoke).
+- Live deployment tests: `pnpm e2e` (one config, `e2e/vitest.config.ts`, two projects). The
+  `node` project runs the engine suites in `e2e/vitest/` (agents, admin-project, preview
+  smoke) and the cross-runtime example matrix in `e2e/examples/`; the `browser` project runs
+  that matrix in a real browser. `pnpm e2e --project node` skips the browser lane.
 - Egress + secret substitution coverage lives in itx suite
-  (`e2e/vitest/itx.e2e.test.ts`); the cross-runtime example matrix is `pnpm e2e:examples`
-  (`e2e/examples/`).
+  (`e2e/vitest/itx.e2e.test.ts`).
 - Preview smoke: `pnpm e2e -t "OS preview smoke"` (`preview-smoke.e2e.test.ts`) exercises a
   deployed preview, including its project MCP route (it derives its project slug from
   `GITHUB_SHA` when set).
